If you don't have a DataFrame called `df_reference` containing the reference data, you won't be able to calculate accuracy, F1 score, and precision. These evaluation metrics require a reference or ground truth data to compare the predicted entities with.
You need to have a labeled dataset where the entities are manually annotated or labeled. This labeled dataset will serve as the reference or ground truth data for evaluating the performance of your entity recognition system.

Once you have the labeled dataset, you can load it into a DataFrame similar to `df_PA4` and then pass both `df_PA4` (predicted data) and `df_reference` (reference data) to the `calculate_metrics` function.

Without the reference data, it is not possible to calculate accuracy, F1 score, and precision accurately. These metrics are used to measure the performance of the model against the known ground truth.